Keep track of users within a group
This change adds the chat.add_user method to track which users are
active in a group. It works if you add users or if they interact with
the bot. It may not be up-to-date if they are added *after* the bot and
they haven't yet done or said anything, but it should fail gracefully in
that case.

This is to allow the bot to ping other users in a group when a user
submits a voicenote. It thanks them based on their user_name if defined,
otherwise their first name, and this seems to work.

Fixes #10