Fixed intervals + textual changes
Added the interval increments so that the different 'days' of the experience are actually triggered on the right interval, counting from the start date. Also made some changes to the instructions, replaced a couple of 'Wispers' with 'Echo', and removed some debug logging of when the next prompt will be sent.